Some clean up
